+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - Issue #1519: Fix initialization of print module.
 - Searchbar and LayerSelection: Behavior restored to the way it was before layerSelection filtering was introduced.
 - 3D: When switching from 2D to 3D, the center coordinate of the current map section can be reproduced in 3D. This can be configured using `map3dParameter.camera.offset`. 
+- Print: Layout A0: WMS image layer are printed correctly.
 
 ---